HistoricalOps.filter overflows when computing the scale factor
Summary: Computing the scale factor may cause an integer overrun for certain values, such as a very large interval. This overrun will cause corruption of the historical data, including access or rejection counts.

The overrun happens because the scale factor is computed using unbounded values instead of the bounded values.

Test: atest FrameworksMockingServicesTests:com.android.server.appop.AppOpsServiceTest

Change-Id: Ia14ceb72e081b7fec1c39cc9c4069f6b00be70a3
Signed-off-by: Serban Ungureanu <serbanu@meta.com>